requirements.txt
Augmentor==0.2.6
gym==0.15.3
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
torch==1.3.1
torchvision==0.4.2
imageio
matplotlib
numpy
opencv-python
pandas
Pillow
pyserial
seaborn
sklearn
stable-baselines
#centrosome
